+++ packages/guide/src/app/locations/components/lab/lab.component.ts
@@ -1,0 +1,38 @@
+import { Component } from '@angular/core';
+import { BaseSelectableContainerComponent } from "../../../shared/components/base-selectable-container/base-selectable-container.component";
+import { ItemProcessShopData, OpeningHours, ShopItemData } from "@ci/data-types";
+import { Observable } from "rxjs";
+import { UiIcon } from "../../../shared/enums/ui-icon.enum";
+
+@Component({
+    selector: 'app-lab',
+    templateUrl: './lab.component.html',
+})
+export class LabComponent extends BaseSelectableContainerComponent<ShopItemData> {
+
+    protected shopItemData$: Observable<ShopItemData[]>;
+    protected uiIcon = UiIcon;
+    protected openingHours$: Observable<Record<string, OpeningHours>>;
+    protected itemProcessing$: Observable<ItemProcessShopData[]>;
+    protected selectedProcessEntity: ItemProcessShopData | undefined;
+    protected showProcessingTable = false
+
+
+    constructor() {
+        super();
+        this.shopItemData$ = this._database.fetchShopItemData$("lab");
+        this.openingHours$ = this._database.fetchOpeningHours$("lab");
+        this.itemProcessing$ = this._database.fetchShopProcessItems$("lab");
+    }
+
+    override showDetails(selectedEntry?: ShopItemData) {
+        this.selectedProcessEntity = undefined;
+        super.showDetails(selectedEntry);
+    }
+
+    showProcessDetails(selectedEntry?: ItemProcessShopData) {
+        this.selectedEntity = undefined;
+        this.selectedProcessEntity = selectedEntry;
+        this.openDrawer = true;
+    }
+}