+++ packages/guide/src/app/locations/locations.module.ts
@@ -12,9 +12,16 @@ import { BlacksmithComponent } from './components/blacksmith/blacksmith.componen
 import { CraftingModule } from '../crafting/crafting.module';
 import { ShopItemDataDetailsComponent } from './components/shop-item-data-details/shop-item-data-details.component';
 import { ShopItemDataTableComponent } from './components/tables/shop-item-data-table/shop-item-data-table.component';
-import {
-    ShopItemProcessTableComponent
-} from './components/tables/shop-item-process-table/shop-item-process-table.component';
+import { ShopItemProcessTableComponent } from './components/tables/shop-item-process-table/shop-item-process-table.component';
+import { LabComponent } from './components/lab/lab.component';
+import { ShopItemProcessDataDetailsComponent } from './components/shop-item-process-data-details/shop-item-process-data-details.component';
+import { OpeningHoursComponent } from './components/opening-hours/opening-hours.component';
+import { GeneralStoreComponent } from './components/general-store/general-store.component';
+import { CarpenterComponent } from './components/carpenter/carpenter.component';
+import { ItemUpgradeDetailsComponent } from './components/item-upgrade-details/item-upgrade-details.component';
+import { ItemUpgradeTableComponent } from './components/tables/item-upgrade-table/item-upgrade-table.component';
+import { MerfolkGeneralStoreComponent } from './components/merfolk-general-store/merfolk-general-store.component';
+import { MerfolkOracleTailStoreComponent } from './components/merfolk-oracle-tail-store/merfolk-oracle-tail-store.component';
 
 @NgModule({
     declarations: [
@@ -25,6 +32,15 @@ import {
         ShopItemDataDetailsComponent,
         ShopItemDataTableComponent,
         ShopItemProcessTableComponent,
+        LabComponent,
+        GeneralStoreComponent,
+        CarpenterComponent,
+        ShopItemProcessDataDetailsComponent,
+        OpeningHoursComponent,
+        ItemUpgradeDetailsComponent,
+        ItemUpgradeTableComponent,
+        MerfolkGeneralStoreComponent,
+        MerfolkOracleTailStoreComponent,
     ],
     imports: [
         CommonModule,
@@ -36,7 +52,7 @@ import {
         MatSortModule,
         CraftingModule,
     ],
-    exports: [OfferingsTableComponent, ShopItemDataTableComponent, ShopItemProcessTableComponent],
+    exports: [OfferingsTableComponent, ShopItemDataTableComponent, ShopItemProcessTableComponent, ItemUpgradeTableComponent],
 })
 export class LocationsModule {
 }